Trim itfMesManager sibling loop locals

+++ src/Event/mt_evtMisc.c
@@ -722,7 +722,7 @@ void FUN_0038a730(int param_1);
 void FUN_0038a7c0(u32 *param_1,u32 *param_2);
 int FUN_0038a920(int param_1,u16 *param_2,int param_3);
 void FUN_0038af70(u16 *param_1);
-void FUN_0038b140(long param_1);
+void FUN_0038b140(int param_1);
 void FUN_0038b2c0(u16 *param_1,u32 *param_2,u32 *param_3,u32 *param_4, int param_5);
 void FUN_0038b550(int param_1,u32 *param_2,u32 *param_3,u32 *param_4);
 void FUN_0038b600(u32 *param_1,u32 *param_2);
@@ -4234,7 +4234,7 @@ void FUN_0038af70(u16 *param_1)
 // FUN_0038B140 NONMATCHING
 
 
-void FUN_0038b140(long param_1)
+void FUN_0038b140(int param_1)
